+++ truffle/com.oracle.truffle.api.interop/src/com/oracle/truffle/api/interop/GenericObjectAccessNode.java
@@ -57,14 +57,14 @@ final class GenericObjectAccessNode extends ObjectAccessNode {
         final ForeignAccess fa = truffleObject.getForeignAccess();
         final CallTarget ct = fa.access(access);
         if (ct == null) {
-            throw messageNotRecognizedException(fa);
+            throw messageNotRecognizedException();
         }
         return ct;
     }
 
     @CompilerDirectives.TruffleBoundary
-    private RuntimeException messageNotRecognizedException(final ForeignAccess fa) {
-        throw new IllegalStateException("Message " + access + " not recognized by " + fa);
+    private RuntimeException messageNotRecognizedException() {
+        throw UnsupportedMessageException.raise(access);
     }
 
 }