+++ truffle/com.oracle.truffle.sl.test/src/com/oracle/truffle/sl/test/SLDebugLegacyTest.java
@@ -65,7 +65,7 @@ import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.interop.ForeignAccess;
 import com.oracle.truffle.api.interop.ForeignAccess.Factory;
-import com.oracle.truffle.api.interop.ForeignAccess.Factory10;
+import com.oracle.truffle.api.interop.ForeignAccess.Factory18;
 import com.oracle.truffle.api.interop.Message;
 import com.oracle.truffle.api.interop.TruffleObject;
 import com.oracle.truffle.api.nodes.RootNode;
@@ -544,7 +544,7 @@ public class SLDebugLegacyTest {
 
     }
 
-    private static class ExecNotifyHandlerForeign implements Factory10, Factory {
+    private static class ExecNotifyHandlerForeign implements Factory18, Factory {
 
         private final ExecNotifyHandler nh;
 
@@ -615,6 +615,11 @@ public class SLDebugLegacyTest {
         @Override
         public boolean canHandle(TruffleObject to) {
             return (to instanceof ExecNotifyHandler);
+        }
+
+        @Override
+        public CallTarget accessKeys() {
+            return null;
         }
 
     }