+++ truffle/com.oracle.truffle.sl/src/com/oracle/truffle/sl/runtime/SLObjectMessageResolution.java
@@ -40,10 +40,12 @@
  */
 package com.oracle.truffle.sl.runtime;
 
+import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.interop.MessageResolution;
 import com.oracle.truffle.api.interop.Resolve;
 import com.oracle.truffle.api.interop.UnknownIdentifierException;
+import com.oracle.truffle.api.interop.java.JavaInterop;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.object.DynamicObject;
 import com.oracle.truffle.sl.SLLanguage;
@@ -120,6 +122,19 @@ public class SLObjectMessageResolution {
             } else {
                 throw UnknownIdentifierException.raise(name);
             }
+        }
+    }
+
+    @Resolve(message = "KEYS")
+    public abstract static class SLForeignPropertiesNode extends Node {
+        public Object access(DynamicObject receiver) {
+            return obtainKeys(receiver);
+        }
+
+        @CompilerDirectives.TruffleBoundary
+        private static Object obtainKeys(DynamicObject receiver) {
+            Object[] keys = receiver.getShape().getKeyList().toArray();
+            return JavaInterop.asTruffleObject(keys);
         }
     }